Add Flower decoration scene
This is a flavour of Decoration which chooses a random texture from a
list. The list in the scene corresponds to the flowers (and a mushroom)
that are currently manually placed in Fray's End, plus a new flower
drawn by a learner (who agrees to license it under CC-BY-SA-4.0, as does
their adult (me)).

The random selection is done based on the node's path in the scene,
which means that the choice will remain the same between runs so long as
the list remains the same.

https://github.com/endlessm/threadbare/issues/556