GameState: Persist the incorporating_threads flag
Add a global section to the state for this flag. Add a setter that not
only changes the flag, but also persists it. Note that this can be a
setter for the current variable because we want to set it without
persisting when initializing and when restoring the saved state.